requirements.txt
numpy
opencv-python
pillow
torch>=1.2.0
torchvision
tqdm
